initial extension setup (#1)
* chore: Add VSCode workspace configuration

* chore: add devDependencies for testing and linting

Add devDependencies for testing with Mocha and VSCode, linting
with ESLint and TypeScript ESLint plugin, and bundling with
webpack. Update lockfile with new dependencies and their
transitive dependencies. Configure to automatically install peer
dependencies.

* ci: add pull request workflow with build, lint and test

Add a new GitHub Actions workflow for pull requests that runs on
macOS, Ubuntu and Windows. The workflow caches pnpm packages,
installs Node.js, builds the project, lints the code and runs tests.

* ci: skip tests on Ubuntu due to $DISPLAY issue

The tests are skipped on the ubuntu-latest runner because of an
issue with the $DISPLAY environment variable. The tests continue
to run on the macos-latest and windows-latest runners.