chore: switch to actions/upload-artifact for extension upload (#54)
* chore: switch to actions/upload-artifact for extension upload

Replace the upload step in build workflow with actions/upload-artifact
to simplify and standardize the process. Comment out deprecated code 
using gh release upload.

* ci: comment out conditional checks in build workflow

Comment out `if: github.ref == 'refs/heads/main'` conditions
for 'Create Pre-release extension' and 'Upload extension'
steps to allow these steps to run on all branches. This is
necessary to enable testing and debugging on non-main branches.

* fix(ci): correct conditional for pre-release extension creation

Update the condition to run the pre-release extension creation
and upload steps only on the 'ubuntu-latest' OS. This ensures
the steps are not mistakenly skipped and aligns with matrix OS
requirements.

* fix(workflows): enable main branch releases and pre-releases

Update release.yml to trigger workflows only on the main branch.
Adjust build.yml to create and upload pre-release extensions
exclusively for the main branch on Ubuntu. These changes ensure
that workflows are correctly scoped to the intended branch.