fix(module-translation): ensure valid path before applying change (#225)
Check for the existence of target values before applying changes
in translation module to prevent invalid / unused translations when using multiple plural forms.

This addresses issues where pluralisation is used in languages with multiple plural forms, next to using languages with less plural forms. This caused the multiple plural forms to be added to the language resource files with less plural forms.