ci: update workflows for selective builds (#269)
* ci: update workflows for selective builds

Add condition to build workflow to skip CI/doc commits.
Expand release workflow to trigger on 'release/**' branches.

* ci: update build condition to use pull request title

Update the build condition to check the pull request title instead of the head commit message. This ensures that the pipeline triggers appropriately on PRs with relevant titles and avoids unnecessary builds for PRs focused on ci or docs-related tasks.