fix: update release process to mark as stable release (#288)
Change preRelease from true to false to promote the current release as a stable version. 
Update the npm extension packaging process to reflect this change. 
This ensures the release is available to all users as a stable version, rather than as a pre-release.

In the future the pre-release will be re-enstated and will be created from the release pull request, ensuring the possibility of testing an actual pre-release.